Refactored BaseOpModes and the location of System to fix broken refactoring.
