+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/VuforiaBaseOpMode.java
@@ -17,7 +17,7 @@ import org.firstinspires.ftc.robotcore.external.matrices.OpenGLMatrix;
  */
 
 
-public abstract class VuforiaBaseOpMode extends BaseOpMode
+public abstract class VuforiaBaseOpMode extends BaseAutonomousOpMode
 {
 
     protected static class Target
