+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/systems/Eye.java
@@ -15,7 +15,7 @@ import org.firstinspires.ftc.robotcore.external.navigation.VuforiaLocalizer;
 import org.firstinspires.ftc.robotcore.external.navigation.VuforiaTrackable;
 import org.firstinspires.ftc.robotcore.external.navigation.VuforiaTrackableDefaultListener;
 import org.firstinspires.ftc.robotcore.external.navigation.VuforiaTrackables;
-import org.firstinspires.ftc.teamcode.robot.systems.System;
+import org.firstinspires.ftc.teamcode.robot.systems.common.System;
 
 /**
  * Created by Steven Abbott on 9/27/2017.
