+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/systems/ParallelLiftSystem.java
@@ -7,6 +7,7 @@ import com.qualcomm.robotcore.hardware.DigitalChannel;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
+import org.firstinspires.ftc.teamcode.robot.systems.common.System;
 
 /**
  * Created by jacks on 11/28/2017.
