+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/systems/PixySystem.java
@@ -6,6 +6,7 @@ import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.Servo;
 
 import org.firstinspires.ftc.teamcode.hardware.pixycam.PixyCam;
+import org.firstinspires.ftc.teamcode.robot.systems.common.System;
 
 @Autonomous(name = "PixySystem", group = "Bot")
 public class PixySystem extends System
