+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/teleop/BaseTeleOpMode.java
@@ -15,7 +15,7 @@ import org.firstinspires.ftc.teamcode.logger.LoggingService;
  * Created by EvanCoulson on 10/11/17.
  */
 
-public abstract class BaseOpMode extends OpMode
+public abstract class BaseTeleOpMode extends OpMode
 {
     protected final ConfigParser config;
     protected Controller controller1;
@@ -27,7 +27,7 @@ public abstract class BaseOpMode extends OpMode
     protected Logger logger;
 
 
-    public BaseOpMode(String opModeName)
+    public BaseTeleOpMode(String opModeName)
     {
         this.logger = new Logger(this, opModeName);
         logger.setLoggingServices(LoggingService.FILE);
